Convert piem-am into a standalone command
b4 is great, and I have no desire to create an Elisp implementation of
its complex patch series extraction.  However, in cases where projects
permit patches as attachments, using b4 isn't an option.  And even for
inline patches, it's useful to be able to git-am the patch associated
with the current buffer.

To support the above cases, teach piem-am to get an mbox from
piem-am-ready-mbox when called interactively.