Add command for downloading/injecting t.mbox.gz into a Maildir
I follow some public-inbox-archived projects over NNTP with Gnus.
However, once I get involved in a thread, I usually want the entire
thread in my local mail/Notmuch database (or at least the message I'm
replying to).

With this new command, if I see a message in Gnus (or EWW or Elfeed)
that I want to reply to, I can get it into Notmuch, assuming
piem-inboxes and piem-maildir-directory are wired up.  And through
piem-after-mail-injection-functions, it's possible to jump straight to
the message with notmuch-show.