Support adding Message-Id to attached patches
With an inline patch that has a Message-Id, this information can be
linked up to the patch when applied (e.g., with git-am's --message-id
flag or using a post-applypatch hook [1]).  Unfortunately, this method
fails for projects where it is common to attach patches, as there is
no Message-Id within the patch.

Teach piem-am-ready-mbox how to insert the Message-Id that piem-mid
reports, which should always correspond to the message that contains
the patch attachments.

[1] Here's an example that used to keep the commit->Message-Id mapping
    in git.git:
    https://lore.kernel.org/git/xmqq7e5ag4g5.fsf@gitster-ct.c.googlers.com/

Message-Id: <20200613044933.4046-1-kyle@kyleam.com>