piem-am: Move git-am arguments to variable
It's tempting to define a transient for piem-am and let that handle
the arguments, but doing so only takes care of the case where piem-am
is the entrypoint, not piem-b4-am---which is already a transient with
b4 arguments.  And it's not yet clear whether many users will need to
modify that these arguments across different calls (e.g., based on
conventions across different projects).  For my purposes, a single set
of arguments is fine.

For now, add a defvar so that there's at least something to latch onto
if someone needs to modify the arguments unconditionally or via a
let-binding.