piem-am: Call git-am with --3way
I think I left this out initially with the idea that the user can (and
should) set am.threeWay to true.  But I can't come up with a reason
why a caller wouldn't want to fall back to a 3-way merge [*], so pass
--3way as an explicit argument to benefit users that haven't
configured am.threeWay.  Now that piem-am-args exists, a user can
always remove it if desired.

[*] And magit-am has had --3way on by default for a long time; I don't
    recall any complaints.