Explicitly specify --patch-format in git-am calls
The sources of mbox patches fed to git-am are 1) threads downloaded
from a public-inbox HTTP instance, 2) mboxes generated via
piem-mid-to-thread-functions, and 3) those generated via
piem-am-ready-mbox-functions.  The first source should always be
mboxrd.  For the second, piem-notmuch-mid-to-thread is currently the
only function suitable for piem-mid-to-thread-functions, and it uses
mboxrd.  The third source is a mix between mbox and mboxrd.

By default, git-am tries to auto-detect the patch format, but let's
explicitly specify --patch-format to avoid any incorrect guesses.

Message-Id: <20200810020704.30150-1-kyle@kyleam.com>