Fix handling of -am-ready-mbox values
0ee97e9 (Explicitly specify --patch-format in git-am calls,
2020-08-09) made it possible for a piem-am-ready-mbox-functions member
to specify the format of the mbox by returning (FUNCTION FORMAT).  If
FUNCTION is returned, then mboxrd is supposed to be taken as the
default format.  The handling is broken, though, because
piem-am-ready-mbox tries to detect the (FUNCTION FORMAT) form with
listp, but that of course also returns true when the return value is
simply a function.

Instead, check to see whether the element matches a valid format
value.  Switch from (FUNCTION FORMAT) to (FUNCTION . FORMAT) to make
it more convenient to pull out FORMAT with cdr-safe.

Message-Id: <877du5c1nz.fsf@kyleam.com>