piem-inbox-coderepo-maybe-read: Include current projectile project
When piem-inbox-coderepo-maybe-read is called from within a project,
the current project isn't included in the collection because
projectile-relevant-known-projects excludes it when
projectile-current-project-on-switch is at its default value.  That's
undesirable in this context; if there's a current project, it's likely
the one of interest (e.g., calling piem-b4-am-ready-from-mbox from a
project's directory).

Add the current project to the collection and make it the default.
Also, don't bother going down the projectile branch if there are no
known projects.

Message-Id: <20200816185130.32703-3-kyle@kyleam.com>