b4: Try to download thread from piem-inboxes URL
If a call to piem-b4-am-from-mid fails to generate the thread for a
message ID via piem-mid-to-thread-functions, b4 is called without a
local mbox.  In this case, b4 tries to download the thread from the
URL specified by b4.midmask in the caller's Git configuration.

That works, but it's inconvenient because the user needs to configure
the URL in two places.  If the current buffer is associated with an
inbox in piem-inboxes, try to download the thread from its :url before
falling back to b4's midmask.

Message-Id: <20200817013343.15615-6-kyle@kyleam.com>