inject: Support skipping messages
piem-inject-thread-into-maildir allows pulling a thread from a
public-inbox URL, but repeating the process to get new messages in the
thread leads to duplicates in the local store.  Add a predicate that
the caller can configure to avoid this.

Note that the predicate has the entire message available to make its
decision, but it's called with the message ID for convenience because
that's likely the only information needed to determine if the message
is already in the local store.

Message-Id: <20200822190130.20397-3-kyle@kyleam.com>