Switch downloads to url-retrieve-synchronously
piem-download-and-decompress uses url-retrieve and a callback, setting
url-asynchronous to nil.  This approach doesn't seem to be sufficient
because I'm getting intermittent failures in piem-b4--get-am-files
related to unfinished processes.  And taking a peek at
url-retrieve-synchronously suggests that indeed a good amount more is
needed to do a synchronous call with url-retrieve.

Switch piem-download-and-decompress over to
url-retrieve-synchronously.  I haven't noticed any issues with the
url-retrieve call in piem-inject-thread-into-maildir, but switch that
over too for simplicity and consistency.

Message-Id: <20200828025605.1106-1-kyle@kyleam.com>