elfeed, eww: Don't assume inbox name in piem-inboxes and URL match
piem-elfeed-get-inbox and piem-eww-get-inbox match the URL against
piem-link-re and take the second group as the inbox name.  That's a
bad approach because the inbox name in the URL doesn't necessarily
match the one in piem-inboxes.  For example, public-inbox's own
archive is https://public-inbox.org/meta/, but my entry in
piem-inboxes uses the name "public-inbox":

    ("public-inbox" :url "https://public-inbox.org/meta/" ...)

The approach also fails if the URL isn't a public-inbox message URL
because piem-link-re isn't very specific.  (That will be improved in
an upcoming commit.)

Find the inbox name by matching the buffer URL against the :url values
in piem-inboxes.

Message-Id: <20200828031920.7515-2-kyle@kyleam.com>