piem-am: Rephrase CODEREPO description
Describing CODEREPO in terms of where git-am is called is a bit
confusing because piem-am does other things here as well (e.g. reading
the base and checking out a branch).  And it won't necessarily be
where git-am is called once worktree support is added.

Give a more generic description.

Message-Id: <20201115061518.22191-2-kyle@kyleam.com>