am: Allow flipping worktree creation with prefix argument
I tend to use a few dedicated worktrees for projects and am not
interested in creating a worktree for each patch series I apply.
However, I can imagine wanting to create one every now and then.  Make
it possible by adding a prefix argument to piem-am and
piem-b4-am-from-mid that flips the meaning of piem-am-create-worktree.

Message-Id: <20201115061518.22191-6-kyle@kyleam.com>