piem-notmuch-am-ready-mbox: Improve handling of attachments
piem-notmuch-am-ready-mbox tries to get attached patches from the
plist returned by notmuch-show-get-message-properties.  This works
okay for simple cases, but it doesn't find a patch if the content is
encoded or if the MIME tree isn't structured as expected.  Instead use
mm-decode functions to get a list of handles and display their
content.

This new approach, as well as the previous one, probably isn't
compatible with `format-patch --attach' patches, but I'm not going to
worry about that until I actually see such a patch in the wild.

Message-Id: <20210104015435.18397-3-kyle@kyleam.com>