piem: Add :maildir keyword to piem-inboxes
Previously the user was able to configure a maildir to inject threads
into, but the user might want the maildir to be different depending
on which mailing list the threads was coming from.

With the `:maildir` keyword, users can configure the maildir on a
per-list basis.  If there is not `:maildir` configured for a mailing
list, it will fallback to the value of `piem-maildir-directory`.

Message-Id: <702dccedfc5e67a41bb0dd58fe66af6e3f204bb5.1615568004.git.public@yoctocell.xyz>