piem-use-magit: Unconditionally set to t
When piem is loaded, piem-use-magit is enabled if Magit has already
been loaded.  This approach is potentially confusing: a user may want
to use Magit, be happy that it seems to just work, and then confused
when it doesn't work in some later session where loading Magit happens
to not be triggered before loading piem.

All the relevant sites have fboundp guards (and those are cheap), so
there's no advantage to disabling this if Magit isn't enabled at load
time.  Set piem-use-magit to t by default.

Message-Id: <20210522203905.16504-3-kyle@kyleam.com>