b4: Check for message ID match when using current buffer's URL
If piem-mid-to-thread-functions fails to generate the thread,
piem-b4--get-am-files tries to download the mbox from the URL
associated with the current buffer.  However, it uses the message ID
returned by piem-mid rather than the message ID passed by
piem-b4-am-from-mid.  That's not a safe assumption for non-interactive
calls to piem-b4-am-from-mid.

Construct the URL with the message ID passed by piem-b4-am-from-mid,
and skip the download completely if that message ID doesn't match the
one for the current buffer.

Message-Id: <20210523214623.31331-2-kyle@kyleam.com>