Rework url-retrieve-synchronously wrapper
piem-download-and-decompress calls url-retrieve-synchronously, checks
for a 200 status, and then manually removes the header.  This works
okay, but it'd be good for the error handling to match what's done by
url-insert-file-contents.

Introduce a new macro that largely copies what is done by
url-insert-file-contents.  The main difference is that url-insert is
used instead of url-insert-buffer-contents so that the contents can be
inserted literally.

This approach is based on Emacs's 5f9671e57e
(lisp/emacs-lisp/package.el: Fix decoding of downloaded files,
2019-05-18).

Message-Id: <20210523214623.31331-3-kyle@kyleam.com>