b4: Better describe how piem-b4-am-ready-from-mid gets the thread
piem-b4-am-ready-from-mid's docstring doesn't mention that it tries to
download the thread from a piem-inboxes URL before falling back to
letting b4 handle the download.  The manual's description is better,
though it makes it sound like the b4 fallback depends on not finding a
URL in piem-inboxes rather than the download being unsuccessful for
whatever reason.  Reword the docstring and manual text to hopefully
make things clearer.

Message-Id: <20210524005040.12668-1-kyle@kyleam.com>