piem-lei-show: Let caller suppress displaying buffer
piem-lei-show switches to the message buffer with pop-to-buffer, but
that behavior won't work well in the context of a mode that gives an
overview of lei-q search results.  In that case, a wrapper command
will want to control the display of the buffer so that it can keep a
split window layout and avoid switching to the piem-lei-show-mode
buffer.

And more generally, Lisp callers are likely to want to handle the
display themselves.  Add an optional 'display' parameter that defaults
to nil for non-interactive calls.

Message-Id: <20210605211402.20304-3-kyle@kyleam.com>