lei: Add command and mode for displaying overview of search results
The output is intended to resemble search in public-inbox's web
interface: an entry for each matching message.  This is different from
notmuch-search's output in that results are not grouped in their
thread.  I like notmuch's interface, although I'm not sure that trying
to reshape lei-q's JSON output into something like that is worth the
code complication or computation cost.

The plan is to eventually wire this up to a transient to allow the
caller to specify arguments (e.g., --only to restrict the search
results to a particular inbox).

Message-Id: <20210605211402.20304-5-kyle@kyleam.com>