lei: Add command for viewing a thread
piem-lei-query presents a message-based overview.  In many cases the
caller will want to use that search result as a seed for finding the
associated thread.  Add a command that construct thread for a given
message.

The threading algorithm is based on public-inbox's.  Some details may
have been lost in translation, but I haven't spotted any differences
yet when doing side-by-side comparisons of output from
piem-lei-query-thread and public-inbox's web interface.  And testing
with a few ~100-message threads, the performance seems to be okay.

The appearance also follows public-inbox's, which I like.

Message-Id: <20210605211402.20304-7-kyle@kyleam.com>