piem-lei-query-thread: Position point on seed message
It seems likely that the caller wants to start digesting the thread in
the context of the seed message, and that message may be part of a
large thread.  Move point to help orient the caller.

Notmuch nicely distinguishes search hits from other messages when
displaying a thread.  Something along those lines is worth considering
eventually.

Message-Id: <20210605211402.20304-9-kyle@kyleam.com>