piem-lei-query-thread: Omit main part of subject if shared
In addition to suppressing identical subjects (after stripping "re:"),
public-inbox's web interface will compare the current line's subject
with the previous line's, and cut off the shared tail:

  [PATCH] Add basic integration for Rmail
  ` <suppressed completely>
    ` [PATCH v2] "                                    <-- here
      ` <suppressed completely>

I think the above is helpful.  However, in some cases, I find the
presentation more confusing than helpful:

  [PATCH 0/3] notmuch: Improve handling of attached patches
  ` [PATCH 1/3] piem-notmuch--with-current-message: Declare debug and indent specs
  ` [PATCH 2/3] piem-notmuch-am-ready-mbox: Improve handling of attachments
    ` <suppressed completely>
      ` [PATCH v2 0/3] notmuch: Improve handling of attached patches
        ` [PATCH v2 1/3] piem-notmuch--with-current-message: Declare debug and indent specs
        ` [PATCH v2 2/3] piem-notmuch-am-ready-mbox: Improve handling of attachments
        ` [PATCH v2 3/3] gnus, notmuch: Absorb now-shared bits into patch attachment helper
  ` [PATCH "

It takes me a second to figure out what the omitted bits in the last
line's subject are.  I'm not sure, but I think the subject truncation
that I find clear is where the omitted text is the main subject after
a bracketed tag (i.e. "[tag] main"), not more or less.

Teach piem-lei-query-thread to split the subject into a "prefix" (some
number of "[tag]" items) and a "main" part (everything else), and
elide a line's main part if it matches the previous line's.  In the
above example, the last line would be

  ` [PATCH 3/3] …

Message-Id: <20210605211402.20304-12-kyle@kyleam.com>