piem-lei-query-thread: Add bug#NNN special case when eliding subject
In debbugs threads, it's not uncommon for a leading "[bug#NNN]" in the
subject to be converted to "bug#NNN:" [*].  I'm not sure what the
source of this is, but it prevents the suppression of an otherwise
identical subject.

It's probably not worth normalizing before the comparison to get full
suppression, but it'd be nice to at least elide the main part of the
subject so it's more obvious that it didn't change.  Add a special
case so that "bug#NNN:" prefix is treated the same as a bracketed
prefix.

[*] example:
    https://yhetil.org/guix-patches/20201128051435.30580-1-kyle@kyleam.com

Message-Id: <20210605211402.20304-13-kyle@kyleam.com>