piem-notmuch-known-mid-p: Drop call to with-current-buffer
There's no need to switch to the standard-output buffer because it can
be passed as a destination for call-process.

I go back and forth on whether I prefer (with-output-to-string ...) to
(with-temp-buffer ... (buffer-string)), but the main thing I like
about with-output-to-string is that the intention of the code is
declared at the start.

Message-Id: <20211025035505.297281-1-kyle@kyleam.com>