lei q: Offer candidates for --include and --only
The --include option of lei-q enables searching external sources that
are not already registered, whether they are local inboxes or remote
URLs.  --only also does this, along with restricting the results to
the specified sources.  As such, registered inboxes make sense as
values for --only, in _addition_ to any values the make sense for
--include.

Message-Id: <20211025035630.297598-10-kyle@kyleam.com>