README: Mention b4 and lei
piem leans on b4 heavily, so it deserves a pointer in the README.
Plus, the next patch will mention b4 when it describes mailscripts,
and it seems a bit odd for that to be the only mention.

And lei wasn't a thing when this README was written, but it's
certainly worth highlighting now.

Message-Id: <20211212222119.124669-2-kyle@kyleam.com>