lei q: Allow initial input for query to be customized
piem-lei-query reads with initial input of "d:20.days.ago.. ", which I
find useful because this time restriction works well with what I'm
most often searching for.  (This is a value I've long used with
Notmuch.)  However, what value of "recent" is most reasonable is
likely to vary across users, and some users may not want the
restriction at all.

Add an option so that the initial input can be easily configured.

Message-Id: <20211228022037.206597-2-kyle@kyleam.com>