lei q: Separate different threads with newline
For piem-lei-query-threads output with multiple threads, add a newline
to make it easier to spot the start of a new thread.

In terms of presentation, I think it would also be nice to distinguish
search hits in the thread from other messages.  For local messages,
pct could be used to do this, but punt on that for now.

Message-Id: <20211228022037.206597-6-kyle@kyleam.com>