lei q: Sort threads by date of initial message
Now that there is a command for displaying multiple threads, the order
that piem-lei-query--thread returns the threads matters.  When lei-q
is given the --threads option, the threads don't have a clear order,
and --sort and --reverse don't have an effect, so
piem-lei-query--thread must handle the sorting.

Teach piem-lei-query--thread to sort by the date of the root message
in a thread, adding a option to control the direction of this sorting.

It may be better to sort by the latest date within the entire thread.
However, doing so is more expensive, so don't bother until a clear
case is made that it's the better option.

Message-Id: <20211228022037.206597-7-kyle@kyleam.com>