lei: Delete message buffer's window along with query window
Within query buffers, a message for the current line can be displayed
while keeping point in the query buffer.  Calling quit-window in a
query buffer, however, leaves the associated buffer's window in place,
which is probably not what the caller wants.  Add a custom quit-window
variant that handles the message buffer too.

Suggested-by: Xinglu Chen <public@yoctocell.xyz>
Link: https://inbox.kyleam.com/piem/871r96am1q.fsf@yoctocell.xyz
Message-Id: <20211228231143.244455-1-kyle@kyleam.com>