piem-am-ready-mbox: Adjust header regex for format-patch change
As of git.git's ba4324c4e1 (e-mail workflow: Message-ID is spelled
with ID in both capital letters, 2023-04-03), git-format-patch inserts
a Message-ID header instead of Message-Id.  Teach
piem--insert-message-id-header to look for either variant when it
checks for an existing header.

Another option would be to ignore the header case entirely.  However,
piem--insert-message-id-header is intended to work only for
format-patch output, so stick with the stricter header matching.

Message-ID: <20230430221552.251335-3-kyle@kyleam.com>