Delete .guix.scm
I found this useful in the beginning, when it had definitions for b4
and piem before they were in Guix.  But I haven't used this in a
while, and it's likely to bit rot if it hasn't already.

I doubt others are depending on it, but they can of course resurrect
it and maintain it locally if they are.