piem-mime--extract-attached-patch: Guard against nil filename
piem-mime--extract-attached-patch considers a handle with type
text/x-diff and text/x-patch to be a patch attachment regardless of
whether there is an associated filename.

The only cases I've found in the wild where a nil file name gets here
are Gnus buffers with emulated MIME (see gnus-article-emulate-mime)
for _inline_ patches.  As of the previous commit, piem-gnus uses the
same logic as piem-notmuch and no longer sends inline patches with
emulated MIME down the attachment code path.

Reported-by: Nicolas Graves <ngraves@ngraves.fr>
Link: https://inbox.kyleam.com/piem/87ikowudpe.fsf@ngraves.fr
Message-ID: <20250303014610.723186-5-kyle@kyleam.com>